(chore) Convert project structure to polymer 2.x element project.
