+++ .editorconfig
@@ -1,0 +1,5 @@
+root = true
+
+[*.{html,json,js,ts}]
+indent_size = 2
+indent_style = space
