+++ bower.json
@@ -1,6 +1,11 @@
 {
   "name": "dolphin-platform-polymer",
   "version": "0.18.0",
+  "description": "The Dolphin Platform framework implements the presentation model pattern and provides a modern way to create enterprise applications.",
+  "keywords": [
+    "Dolphin Platform"
+  ],
+  "homepage": "http://www.dolphin-platform.io/",
   "authors": [
     "Michael Heinrichs <michael.heinrichs@canoo.com>",
     "Hendrik Ebbers <hendrik.ebbers@canoo.com>",
@@ -8,37 +13,25 @@
     "Timo Brandstätter <timo.brandstaetter@canoo.com>",
     "Benjamin Gudehus <benjamin.gudehus@canoo.com>"
   ],
-  "homepage": "http://www.dolphin-platform.io",
-  "description": "The Dolphin Platform framework implements the presentation model pattern and provides a modern way to create enterprise applications.",
-  "keywords": [
-    "Dolphin Platform"
-  ],
   "licenses": [
     {
       "type": "Apache-2.0",
       "url": "http://www.apache.org/licenses/LICENSE-2.0"
     }
   ],
-  "main": "./dist/dolphin-platform-polymer.js",
   "ignore": [
     "**/.*",
-    "lib",
-    "src",
     "test",
-    "gulpfile.js",
-    "package.json",
-    "sauce.launchers.js",
-    "sauce-template.json",
-    "wct.conf.js",
-    "sonar-project.properties"
+    "package.json"
   ],
   "devDependencies": {
-    "core.js": "2.4.0",
-    "dolphin-platform-js": "0.18.0",
-    "web-component-tester": "^6.1.5"
+    "lodash": "3.10.1",
+    "web-component-tester": "6.2.0"
   },
   "dependencies": {
-    "polymer": "^2.0.2",
-    "webcomponentsjs": "webcomponents/webcomponentsjs#^1.0.10"
+    "core.js": "2.5.1",
+    "dolphin-platform-js": "0.18.0",
+    "polymer": "2.1.1",
+    "webcomponentsjs": "webcomponents/webcomponentsjs#1.0.14"
   }
 }
