--- demo/index.html
@@ -1,32 +1,0 @@
-<html>
-<head>
-    <meta charset="utf-8">
-    <meta name="description" content="">
-    <script src="bower_components/webcomponentsjs/webcomponents.js"></script>
-    <link rel="import" href="bower_components/paper-styles/paper-styles.html">
-    <link rel="import" href="bower_components/font-roboto/roboto.html">
-    <script src="../dist/dolphin-platform-polymer.js"></script>
-    <style>
-        body {
-            font-family: 'Roboto', 'Noto', sans-serif;
-            font-size: 14px;
-            margin: 0;
-            padding: 24px;
-        }
-    </style>
-    <script>
-
-        clientContext = dolphin.clientContext('http://localhost:8080/todo-app/dolphin');
-        clientContext.onConnect().then(function() {
-            console.log('Dolphin connected...');
-        }, function(error){
-            console.log('error', error);
-        });
-
-    </script>
-    <link href="todo-view.html" rel="import">
-</head>
-<body unresolved class="fullbleed">
-<todo-view></todo-view>
-</body>
-</html>
