--- demo/todo-view.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="import" href="bower_components/paper-input/paper-input.html">
-    <link rel="import" href="bower_components/paper-button/paper-button.html">
-    <link rel="import" href="bower_components/paper-item/paper-item.html">
-    <link rel="import" href="bower_components/paper-item/paper-item-body.html">
-    <link rel="import" href="bower_components/iron-icon/iron-icon.html">
-    <link rel="import" href="bower_components/iron-icons/iron-icons.html">
-    <link rel="import" href="bower_components/paper-icon-button/paper-icon-button.html">
-</head>
-<dom-module id="todo-view">
-    <style>
-        paper-button.colorful {
-            color: #4285f4;
-        }
-        paper-button[raised].colorful {
-            background: #4285f4;
-            color: #fff;
-        }
-        .input-form {
-            display: flex;
-            flex-direction: row;
-            flex-wrap: nowrap;
-            justify-content: space-between;
-        }
-        .text-input {
-            max-width: 100%;
-            flex-grow: 1;
-        }
-    </style>
-    <template>
-        <div class="layout center-justified">
-
-            <div class="input-form">
-                <paper-input class="text-input" label="Task name" value="{{model.newItemText}}"></paper-input>
-                <paper-button class="colorful" on-tap="_onAddClicked"><iron-icon icon="add"></iron-icon></paper-button>
-            </div>
-            <template is="dom-repeat" items="{{model.items}}">
-                <paper-item>
-                    <template is="dom-if" if="[[item.completed]]">
-                        <iron-icon icon="check"></iron-icon>
-                    </template>
-                    <paper-item-body on-tap="_onCheckboxClicked">[[item.text]]</paper-item-body>
-                    <paper-icon-button icon="clear" on-tap="_onDeleteClicked"></paper-icon-button>
-                </paper-item>
-            </template>
-        </div>
-    </template>
-</dom-module>
-
-<script>
-    Polymer({
-        is: "todo-view",
-
-        behaviors: [clientContext.createBehavior('ToDoController')],
-
-        properties: {
-            clientContext: Object
-        },
-
-        _onAddClicked: function() {
-            this.invoke('add');
-        },
-
-        _onCheckboxClicked: function(e) {
-            this.invoke('change', {item: e.model.item.text});
-        },
-
-        _onDeleteClicked: function(e) {
-            this.invoke('remove', {item: e.model.item.text});
-        }
-    });
-</script>
