+++ package.json
@@ -2,44 +2,17 @@
   "name": "dolphin-platform-polymer",
   "version": "0.18.0",
   "description": "Polymer Client of the Dolphin Platform",
-  "main": "./dist/dolphin-platform-polymer.js",
+  "author": "Michael Heinrichs <michael.heinrichs@canoo.com>",
   "repository": {
     "type": "git",
     "url": "https://github.com/canoo/dolphin-platform-polymer.git"
   },
-  "author": "Michael Heinrichs <michael.heinrichs@canoo.com>",
   "license": "Apache-2.0",
   "scripts": {
-    "test": "./node_modules/.bin/gulp verify"
+    "test": "./node_modules/.bin/wct --local chrome"
   },
   "devDependencies": {
-    "bower": "1.8.x",
-    "browserify": "13.0.x",
-    "chai": "3.5.x",
-    "del": "2.2.x",
-    "glob": "7.0.x",
-    "gulp": "3.9.x",
-    "gulp-derequire": "2.1.x",
-    "gulp-jshint": "2.0.x",
-    "gulp-load-plugins": "1.2.x",
-    "gulp-mocha": "2.2.x",
-    "gulp-rename": "1.2.x",
-    "gulp-sourcemaps": "1.6.x",
-    "gulp-uglify": "1.5.x",
-    "gulp-util": "3.0.x",
-    "jshint": "2.9.x",
-    "lodash": "4.12.x",
-    "lodash.assign": "4.0.x",
-    "mocha": "2.4.x",
-    "sinon": "1.17.x",
-    "vinyl-buffer": "1.0.x",
-    "vinyl-source-stream": "1.1.x",
-    "watchify": "3.7.x",
-    "web-component-tester": "^6.1.5",
-    "web-component-tester-istanbul": "^0.10.0"
-  },
-  "dependencies": {
-    "@types/polymer": "^1.2.1",
-    "@types/webcomponents.js": "^0.6.32"
+    "bower": "1.8.2",
+    "web-component-tester": "6.2.0"
   }
 }
