--- sauce.launchers.js
@@ -1,122 +1,0 @@
-exports.browsers = [
-    {
-        platform: 'Windows 7',
-        browserName: 'internet explorer',
-        version: '10.0'
-    },
-    {
-        platform: 'Windows 7',
-        browserName: 'internet explorer',
-        version: '11.0'
-    },
-    {
-        platform: 'Windows 7',
-        browserName: 'chrome'
-    },
-    {
-        platform: 'Windows 7',
-        browserName: 'firefox'
-    },
-
-    {
-        platform: 'Windows 8',
-        browserName: 'internet explorer',
-        version: '10.0'
-    },
-    {
-        platform: 'Windows 8',
-        browserName: 'chrome'
-    },
-    {
-        platform: 'Windows 8',
-        browserName: 'firefox'
-    },
-
-    {
-        platform: 'Windows 8.1',
-        browserName: 'internet explorer',
-        version: '11.0'
-    },
-    {
-        platform: 'Windows 8.1',
-        browserName: 'chrome'
-    },
-    {
-        platform: 'Windows 8.1',
-        browserName: 'firefox'
-    }
-    ,
-    // {
-    //     browserName: "safari",
-    //     platform: "OS X 10.11",
-    //     version: "9"
-    // },
-    // {
-    //     platformName: 'iOS',
-    //     platformVersion: '8.4',
-    //     deviceName: 'iPad Simulator',
-    //     browserName: 'Safari',
-    //     deviceOrientation: 'landscape'
-    // },
-    // {
-    //     platformName: 'iOS',
-    //     platformVersion: '9.2',
-    //     deviceName: 'iPad Simulator',
-    //     browserName: 'Safari',
-    //     deviceOrientation: 'landscape'
-    // },
-    // {
-    //     platformName: 'iOS',
-    //     platformVersion: '9.3',
-    //     deviceName: 'iPad Simulator',
-    //     browserName: 'Safari',
-    //     deviceOrientation: 'landscape'
-    // },
-
-    {
-        platform: 'Linux',
-        browserName: 'chrome'
-    },
-    {
-        platform: 'Linux',
-        browserName: 'firefox'
-    },
-
-    {
-        platform: 'OS X 10.10',
-        browserName: 'chrome'
-    },
-    {
-        platform: 'OS X 10.10',
-        browserName: 'firefox'
-    }
-    ,
-    // {
-    //     platform: 'OS X 10.10',
-    //     browserName: 'safari'
-    // },
-
-    {
-        platform: 'OS X 10.11',
-        browserName: 'chrome'
-    },
-    {
-        platform: 'OS X 10.11',
-        browserName: 'firefox'
-    },
-    // {
-    //     platform: 'OS X 10.11',
-    //     browserName: 'safari'
-    // },
-
-    {
-        platform: 'Windows 7',
-        browserName: 'internet explorer',
-        version: '10.0'
-    },
-    {
-        platform: 'Windows 8',
-        browserName: 'internet explorer',
-        version: '10.0'
-    }
-];
