--- sonar-project.properties
@@ -1,23 +1,0 @@
-sonar.projectKey=dolphin-platform-polymer
-sonar.projectName=Polymer client API of the Dolphin Platform (dolphin-platform-Polymer)
-sonar.projectVersion=0.18.0
-sonar.branch=master
-
-# =====================================================
-#   Meta-data for the project
-# =====================================================
-
-sonar.links.homepage=https://github.com/canoo/dolphin-platform-polymer
-sonar.links.ci=https://travis-ci.org/canoo/dolphin-platform-polymer
-sonar.links.scm=https://github.com/canoo/dolphin-platform-polymer
-
-# SQ standard properties
-sonar.language=js
-sonar.sourceEncoding=UTF-8
-sonar.sources=src
-sonar.tests=test/src
-
-# Properties specific to language plugins:
-#   - For JavaScript
-#todo generate lcov report from karma/jasmine test
-sonar.javascript.lcov.reportPath=coverage/lcov.info
