+++ src/binder.js
@@ -13,146 +13,143 @@
  * limitations under the License.
  */
 
-/*jslint browserify: true */
 "use strict";
 
-var Map  = require('../bower_components/core.js/library/fn/map');
-
-
+var Map = require('../bower_components/core.js/library/fn/map');
 
 function exists(object) {
-    return typeof object !== 'undefined' && object !== null;
+  return typeof object !== 'undefined' && object !== null;
 }
 
 function bindScope(scope, fn) {
-    return function () {
-        fn.apply(scope, arguments);
-    };
+  return function () {
+    fn.apply(scope, arguments);
+  };
 }
 
 function deepEqual(array1, array2) {
-    if (array1 === array2 || (!exists(array1) && !exists(array2))) {
-        return true;
-    }
-    if (exists(array1) !== exists(array2)) {
-        return false;
-    }
-    var n = array1.length;
-    if (array2.length !== n) {
-        return false;
-    }
-    for (var i = 0; i < n; i++) {
-        if (array1[i] !== array2[i]) {
-            return false;
-        }
-    }
+  if (array1 === array2 || (!exists(array1) && !exists(array2))) {
     return true;
+  }
+  if (exists(array1) !== exists(array2)) {
+    return false;
+  }
+  var n = array1.length;
+  if (array2.length !== n) {
+    return false;
+  }
+  for (var i = 0; i < n; i++) {
+    if (array1[i] !== array2[i]) {
+      return false;
+    }
+  }
+  return true;
 }
-
 
 function Binder(beanManager) {
-    this.listeners = new Map();
+  this.listeners = new Map();
 
-    beanManager.onBeanUpdate(bindScope(this, this.onBeanUpdateHandler));
-    beanManager.onArrayUpdate(bindScope(this, this.onArrayUpdateHandler));
+  beanManager.onBeanUpdate(bindScope(this, this.onBeanUpdateHandler));
+  beanManager.onArrayUpdate(bindScope(this, this.onArrayUpdateHandler));
 }
 
-
-Binder.prototype.onBeanUpdateHandler = function(bean, propertyName, newValue, oldValue) {
-    if (oldValue === newValue) {
-        return;
-    }
-    var listenerList = this.listeners.get(bean);
-    if (exists(listenerList) && listenerList.length > 0) {
-        var entry = listenerList[0];
-        var element = entry.element;
-        var path = entry.rootPath + '.' + propertyName;
-        element.set(path, newValue);
-    } else {
-        bean[propertyName] = newValue;
-    }
+Binder.prototype.onBeanUpdateHandler = function (bean, propertyName, newValue, oldValue) {
+  if (oldValue === newValue) {
+    return;
+  }
+  var listenerList = this.listeners.get(bean);
+  if (exists(listenerList) && listenerList.length > 0) {
+    var entry = listenerList[0];
+    var element = entry.element;
+    var path = entry.rootPath + '.' + propertyName;
+    element.set(path, newValue);
+  }
+  else {
+    bean[propertyName] = newValue;
+  }
 };
 
-
-Binder.prototype.onArrayUpdateHandler = function(bean, propertyName, index, count, newElements) {
-    var array = bean[propertyName];
-    var oldElements = array.slice(index, index + count);
-    if (deepEqual(newElements, oldElements)) {
-        return;
+Binder.prototype.onArrayUpdateHandler = function (bean, propertyName, index, count, newElements) {
+  var array = bean[propertyName];
+  var oldElements = array.slice(index, index + count);
+  if (deepEqual(newElements, oldElements)) {
+    return;
+  }
+  var listenerList = this.listeners.get(bean);
+  if (exists(listenerList) && listenerList.length > 0) {
+    var entry = listenerList[0];
+    var element = entry.element;
+    var path = entry.rootPath + '.' + propertyName;
+    if (typeof newElements === 'undefined') {
+      element.splice(path, index, count);
     }
-    var listenerList = this.listeners.get(bean);
-    if (exists(listenerList) && listenerList.length > 0) {
-        var entry = listenerList[0];
-        var element = entry.element;
-        var path = entry.rootPath + '.' + propertyName;
-        if (typeof newElements === 'undefined') {
-            element.splice(path, index, count);
-        } else {
-            element.splice.apply(element, [path, index, count].concat(newElements));
-        }
-    } else {
-        if (typeof newElements === 'undefined') {
-            array.splice(index, count);
-        } else {
-            array.splice.apply(array, [index, count].concat(newElements));
-        }
+    else {
+      element.splice.apply(element, [path, index, count].concat(newElements));
     }
+  }
+  else {
+    if (typeof newElements === 'undefined') {
+      array.splice(index, count);
+    }
+    else {
+      array.splice.apply(array, [index, count].concat(newElements));
+    }
+  }
 };
-
 
 Binder.prototype.bind = function (element, rootPath, value) {
-    if (!exists(value) || typeof value !== 'object') {
-        return;
-    }
-    var listenerList = this.listeners.get(value);
-    if (!exists(listenerList)) {
-        listenerList = [];
-        this.listeners.set(value, listenerList);
-    }
-    listenerList.push({element: element, rootPath: rootPath});
+  if (!exists(value) || typeof value !== 'object') {
+    return;
+  }
+  var listenerList = this.listeners.get(value);
+  if (!exists(listenerList)) {
+    listenerList = [];
+    this.listeners.set(value, listenerList);
+  }
+  listenerList.push({ element: element, rootPath: rootPath });
 
-    if (Array.isArray(value)) {
-        for (var i = 0; i < value.length; i++) {
-            this.bind(element, rootPath + '.' + i, value[i]);
-        }
-    } else if (typeof value === 'object') {
-        for (var propertyName in value) {
-            if (value.hasOwnProperty(propertyName)) {
-                this.bind(element, rootPath + '.' + propertyName, value[propertyName]);
-            }
-        }
+  if (Array.isArray(value)) {
+    for (var i = 0; i < value.length; i++) {
+      this.bind(element, rootPath + '.' + i, value[i]);
     }
+  }
+  else if (typeof value === 'object') {
+    for (var propertyName in value) {
+      if (value.hasOwnProperty(propertyName)) {
+        this.bind(element, rootPath + '.' + propertyName, value[propertyName]);
+      }
+    }
+  }
 };
 
 Binder.prototype.unbind = function (element, rootPath, value) {
-    if (!exists(value) || typeof value !== 'object') {
-        return;
-    }
-    var listenerList = this.listeners.get(value);
-    if (exists(listenerList)) {
-        var n = listenerList.length;
-        for (var i = 0; i < n; i++) {
-            var entry = listenerList[i];
-            if (entry.element === element && entry.rootPath === rootPath) {
-                listenerList.splice(i, 1);
+  if (!exists(value) || typeof value !== 'object') {
+    return;
+  }
+  var listenerList = this.listeners.get(value);
+  if (exists(listenerList)) {
+    var n = listenerList.length;
+    for (var i = 0; i < n; i++) {
+      var entry = listenerList[i];
+      if (entry.element === element && entry.rootPath === rootPath) {
+        listenerList.splice(i, 1);
 
-                if (Array.isArray(value)) {
-                    for (var j = 0; j < value.length; j++) {
-                        this.unbind(element, rootPath + '.' + j, value[j]);
-                    }
-                } else if (typeof value === 'object') {
-                    for (var propertyName in value) {
-                        if (value.hasOwnProperty(propertyName)) {
-                            this.unbind(element, rootPath + '.' + propertyName, value[propertyName]);
-                        }
-                    }
-                }
-                return;
-            }
+        if (Array.isArray(value)) {
+          for (var j = 0; j < value.length; j++) {
+            this.unbind(element, rootPath + '.' + j, value[j]);
+          }
         }
+        else if (typeof value === 'object') {
+          for (var propertyName in value) {
+            if (value.hasOwnProperty(propertyName)) {
+              this.unbind(element, rootPath + '.' + propertyName, value[propertyName]);
+            }
+          }
+        }
+        return;
+      }
     }
+  }
 };
-
-
 
 exports.Binder = Binder;
