+++ src/dolphin-polymer-api.js
@@ -13,16 +13,14 @@
  * limitations under the License.
  */
 
-/*jslint browserify: true */
 "use strict";
 
 var dolphinClient = require('../bower_components/dolphin-platform-js/dist/dolphin-platform.js');
 var setupCreateBehavior = require('./behavior.js').setupCreateBehavior;
 
-exports.clientContext = function(url, config){
-    var clientContextFactory = new dolphinClient.ClientContextFactory();
-    var clientContext = clientContextFactory.create(url, config);
-    clientContext.createBehavior = setupCreateBehavior(clientContext);
-    return clientContext;
+exports.clientContext = function (url, config) {
+  var clientContextFactory = new dolphinClient.ClientContextFactory();
+  var clientContext = clientContextFactory.create(url, config);
+  clientContext.createBehavior = setupCreateBehavior(clientContext);
+  return clientContext;
 };
-
