+++ test/dolphin-controller.test.html
@@ -1,0 +1,27 @@
+<link rel="import" href="../dolphin-controller.html">
+
+<test-fixture id="simple-fixture">
+  <template>
+    <dolphin-controller></dolphin-controller>
+  </template>
+</test-fixture>
+
+<script>
+  suite("<dolphin-controller>", () => {
+
+    test("simple use case", done => {
+      const element = fixture("simple-fixture")
+
+      flush(() => {
+        element.set("model", { bean: "foo" })
+        element.set("model.bean", "bar")
+        element.set("model.array", [])
+        element.set("model.array", [1, 2, 3, 4])
+        element.splice("model.array", 1, 2, 9)
+        console.log(element.model)
+        done()
+      })
+    })
+
+  })
+</script>
