+++ test/index.html
@@ -1,0 +1,8 @@
+<script src="../bower_components/webcomponentsjs/webcomponents-loader.js"></script>
+<script src="../bower_components/web-component-tester/browser.js"></script>
+
+<script>
+  WCT.loadSuites([
+    "./dolphin-controller.test.html"
+  ])
+</script>
