--- test/standalone_runner.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>dolphin-js Polymer tests</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-    <link rel="stylesheet" href="../node_modules/mocha/mocha.css" />
-    <script src="../bower_components/webcomponentsjs/webcomponents-lite.js"></script>
-    <link rel="import" href="../bower_components/polymer/polymer.html">
-</head>
-<body>
-    <div id="mocha"></div>
-    <script src="../node_modules/mocha/mocha.js"></script>
-    <script src="../node_modules/chai/chai.js"></script>
-    <script src="../node_modules/sinon/pkg/sinon.js"></script>
-    <script>mocha.setup('bdd')</script>
-    <script src="./build/test-bundle.js"></script>
-    <script>
-        expect = chai.expect;
-        window.addEventListener('WebComponentsReady', function() {
-            mocha.run();
-        });
-    </script>
-</body>
-</html>
