--- test/wct_runner.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <script src="../bower_components/webcomponentsjs/webcomponents-lite.js"></script>
-    <script src="../node_modules/web-component-tester/browser.js"></script>
-    <link rel="import" href="../bower_components/polymer/polymer.html">
-</head>
-<body>
-<script src="./build/test-bundle.js"></script>
-</body>
-</html>
